# [HUDI-4221] Optimzing getAllPartitionPaths 