## What is the purpose of the pull request

- We reverted the logic for getAllPartitionPaths in https://github.com/apache/hudi/pull/5829. But based on feedback, made some minor tweaks to the reverted code (leverage spark par for dir processing) and perf numbers are still holding good and infact better in some cases. 

Perf numbers from S3 listing: 

Listing a hudi table w/ 10k partitons with each partition having just 1 file. 
// Each value is avg from 10 runs. 

0.9.0: 11 to 12 secs. 
0.11.0: 120 to 140 secs.
master: 11 to 12 secs. 
this patch: 8 to 9 secs. 

Listing a hudi table w/ 40 partitions w/ each having ~ 1000 files. 

0.9.0: 900 to 1000 msec.
0.11.0: 700 to 850 ms
master: 930 to 1030 ms.
this patch: 200 to 240 ms. 

## Brief change log

- Optimizing getAllPartitionPaths 

## Verify this pull request

Already covered by existing tests.
Ran perf tests over S3. 

## Committer checklist

 - [ ] Has a corresponding JIRA in PR title & commit
 
 - [ ] Commit message is descriptive of the change
 
 - [ ] CI is green

 - [ ] Necessary doc changes done or have another open PR
       
 - [ ] For large changes, please consider breaking it into sub-tasks under an umbrella JIRA.